requirements.txt
fastapi
python-dotenv
pymupdf
pinecone-client
langchain-core
langchain-community
langchain
langchain-deepseek
langgraph